functions/youtube/requirements.txt
google-api-python-client==2.118.0
google-auth==2.28.1
pyyaml
